requirements.txt
python-telegram-bot==20.7
yt-dlp
ffmpeg-python
google-auth-oauthlib
google-auth-httplib2
google-api-python-client